Step3 피드백적용
