+++ Pepero/main.swift
@@ -7,26 +7,6 @@
 
 import Foundation
 
-func makePepero(length: Int, body: String, topping: String, stickLength: Int) {
-    print("""
-    <정보>
-    길이: \(length)
-    몸통: \(body)
-    토핑: \(topping)
-    막대길이: \(stickLength)\n
-    """)
-    
-    for _ in 1...length {
-        print(topping, body, topping, separator: "")
-    }
-    
-    for _ in 1...stickLength {
-        print(" | |")
-    }
-    
-    print("\n")
-}
-
 makePepero(length: 10, body: "***", topping: " ", stickLength: 4)
 makePepero(length: 12, body: "***", topping: "&", stickLength: 4)
 makePepero(length: 12, body: "***", topping: "#", stickLength: 6)
